+++ presentation/src/main/kotlin/land/sungbin/androidprojecttemplate/common/component/LoadingIndicator.kt
@@ -1,0 +1,18 @@
+package land.sungbin.androidprojecttemplate.common.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun DuckieLoadingIndicator(){
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ){
+        CircularProgressIndicator()
+    }
+}
